+++ micro_manager/adaptivity/global_adaptivity.py
@@ -26,6 +26,7 @@ class GlobalAdaptivityCalculator(AdaptivityCalculator):
         base_logger,
         rank: int,
         comm,
+        micro_problem_cls,
     ) -> None:
         """
         Class constructor.
@@ -46,8 +47,12 @@ class GlobalAdaptivityCalculator(AdaptivityCalculator):
             MPI rank.
         comm : MPI.Comm
             Communicator for MPI.
+        micro_problem_cls : callable
+            Class of micro problem.
         """
-        super().__init__(configurator, global_number_of_sims, base_logger, rank)
+        super().__init__(
+            configurator, global_number_of_sims, micro_problem_cls, base_logger, rank
+        )
         self._global_number_of_sims = global_number_of_sims
         self._global_ids = global_ids
         self._comm = comm
@@ -455,9 +460,7 @@ class GlobalAdaptivityCalculator(AdaptivityCalculator):
         # Only handle activation of simulations on this rank
         for gid in to_be_activated_gids:
             to_be_activated_lid = self._global_ids.index(gid)
-            micro_sims[to_be_activated_lid] = create_simulation_class(
-                self._micro_problem
-            )(gid)
+            micro_sims[to_be_activated_lid] = self._micro_problem_cls(gid)
             assoc_active_gid = self._sim_is_associated_to[gid]
 
             if self._is_sim_on_this_rank[
@@ -494,9 +497,7 @@ class GlobalAdaptivityCalculator(AdaptivityCalculator):
             local_ids = to_be_activated_map[gid]
             for lid in local_ids:
                 # Create the micro simulation object and set its state
-                micro_sims[lid] = create_simulation_class(self._micro_problem)(
-                    self._global_ids[lid]
-                )
+                micro_sims[lid] = self._micro_problem_cls(self._global_ids[lid])
                 micro_sims[lid].set_state(state)
 
         # Delete the micro simulation object if it is inactive