+++ micro_manager/adaptivity/local_adaptivity.py
@@ -13,7 +13,15 @@ from ..micro_simulation import create_simulation_class
 
 
 class LocalAdaptivityCalculator(AdaptivityCalculator):
-    def __init__(self, configurator, num_sims, base_logger, rank, comm) -> None:
+    def __init__(
+        self,
+        configurator,
+        num_sims,
+        base_logger,
+        rank,
+        comm,
+        micro_problem_cls,
+    ) -> None:
         """
         Class constructor.
 
@@ -29,8 +37,10 @@ class LocalAdaptivityCalculator(AdaptivityCalculator):
             Rank of the current MPI process.
         comm : MPI.Comm
             Communicator for MPI.
+        micro_problem_cls : callable
+            Class of micro problem.
         """
-        super().__init__(configurator, num_sims, base_logger, rank)
+        super().__init__(configurator, num_sims, micro_problem_cls, base_logger, rank)
         self._comm = comm
 
         # similarity_dists: 2D array having similarity distances between each micro simulation pair
@@ -283,7 +293,7 @@ class LocalAdaptivityCalculator(AdaptivityCalculator):
         # Update the set of inactive micro sims
         for i in to_be_activated_ids:
             associated_active_id = self._sim_is_associated_to[i]
-            micro_sims[i] = create_simulation_class(self._micro_problem)(i)
+            micro_sims[i] = self._micro_problem_cls(i)
             micro_sims[i].set_state(micro_sims[associated_active_id].get_state())
             self._sim_is_associated_to[
                 i