+++ micro_manager/micro_manager.py
@@ -21,6 +21,7 @@ import numpy as np
 from psutil import Process
 import csv
 import subprocess
+from functools import partial
 
 import precice
 
@@ -29,6 +30,7 @@ from .micro_manager_base import MicroManager
 from .adaptivity.global_adaptivity import GlobalAdaptivityCalculator
 from .adaptivity.local_adaptivity import LocalAdaptivityCalculator
 from .adaptivity.global_adaptivity_lb import GlobalAdaptivityLBCalculator
+from .adaptivity.model_adaptivity import ModelAdaptivity
 
 from .domain_decomposition import DomainDecomposer
 
@@ -57,6 +59,7 @@ class MicroManagerCoupling(MicroManager):
         """
         super().__init__(config_file)
 
+        self._log_file = log_file
         self._logger = Logger(__name__, log_file, self._rank)
 
         self._config.set_logger(self._logger)
@@ -137,6 +140,8 @@ class MicroManagerCoupling(MicroManager):
         self._adaptivity_n = self._config.get_adaptivity_n()
         self._adaptivity_output_n = self._config.get_adaptivity_output_n()
 
+        self._is_model_adaptivity_on = self._config.turn_on_model_adaptivity()
+
         # Define the preCICE Participant
         self._participant = precice.Participant(
             "Micro-Manager",
@@ -144,6 +149,9 @@ class MicroManagerCoupling(MicroManager):
             self._rank,
             self._size,
         )
+
+        self._t = 0  # global time
+        self._n = 0  # sim-step
 
     # **************
     # Public methods
@@ -156,7 +164,7 @@ class MicroManagerCoupling(MicroManager):
         - Read data from preCICE, solve micro simulations, and write data to preCICE
         - If adaptivity is on, compute micro simulations adaptively.
         """
-        t, n = 0, 0
+        self._t = self._n = 0
         sim_states_cp = [None] * self._local_number_of_sims
         mem_usage: list = []
         mem_usage_n = []
@@ -164,6 +172,9 @@ class MicroManagerCoupling(MicroManager):
         process = Process()
 
         micro_sim_solve = self._get_solve_variant()
+        # TODO adapt variant to also use special mada case, iterates in itself and will prob
+        # call _solve_micro_simulations or _solve_micro_simulations_with_adaptivity internally
+        # should use ModelAdaptivity methods to coordinate
 
         dt = min(self._participant.get_max_time_step_size(), self._micro_dt)
 
@@ -172,7 +183,7 @@ class MicroManagerCoupling(MicroManager):
 
         if self._is_adaptivity_on:
             # Log initial adaptivity metrics
-            self._adaptivity_controller.log_metrics(n)
+            self._adaptivity_controller.log_metrics(self._n)
 
         while self._participant.is_coupling_ongoing():
 
@@ -180,7 +191,7 @@ class MicroManagerCoupling(MicroManager):
 
             if self._is_adaptivity_on:
                 if (self._adaptivity_in_every_implicit_step or first_iteration) and (
-                    n % self._adaptivity_n == 0
+                    self._n % self._adaptivity_n == 0
                 ):
                     self._participant.start_profiling_section(
                         "micro_manager.solve.adaptivity_computation"
@@ -209,7 +220,7 @@ class MicroManagerCoupling(MicroManager):
 
             if self._is_adaptivity_with_load_balancing:
                 if (
-                    n % self._load_balancing_n == 0
+                    self._n % self._load_balancing_n == 0
                     and (not first_time_window)
                     and first_iteration
                 ):
@@ -294,25 +305,29 @@ class MicroManagerCoupling(MicroManager):
             if (
                 self._participant.is_time_window_complete()
             ):  # Time window has converged, now micro output can be generated
-                t += dt  # Update time to the end of the time window
-                n += 1  # Update time step to the end of the time window
+                self._t += dt  # Update time to the end of the time window
+                self._n += 1  # Update time step to the end of the time window
 
                 if self._micro_sims_have_output:
-                    if n % self._micro_n_out == 0:
+                    if self._n % self._micro_n_out == 0:
                         for sim in self._micro_sims:
                             if sim:
                                 sim.output()
 
-                if self._is_adaptivity_on and (n % self._adaptivity_output_n == 0):
-                    self._adaptivity_controller.log_metrics(n)
+                if self._is_adaptivity_on and (
+                    self._n % self._adaptivity_output_n == 0
+                ):
+                    self._adaptivity_controller.log_metrics(self._n)
 
                 if self._memory_usage_output_type and (
-                    n % self._memory_usage_output_n == 0 or n == 1
+                    self._n % self._memory_usage_output_n == 0 or self._n == 1
                 ):
                     mem_usage.append(process.memory_info().rss / 1024**2)
-                    mem_usage_n.append(n)
+                    mem_usage_n.append(self._n)
 
-                self._logger.log_info_rank_zero("Time window {} converged.".format(n))
+                self._logger.log_info_rank_zero(
+                    "Time window {} converged.".format(self._n)
+                )
                 first_iteration = (
                     True  # Reset first iteration flag for the next time window
                 )
@@ -321,13 +336,16 @@ class MicroManagerCoupling(MicroManager):
                 first_iteration = True
 
         # Final memory usage logging at the end of the simulation if not already logged at the end of the last time window
-        if self._memory_usage_output_type and n % self._memory_usage_output_n != 0:
+        if (
+            self._memory_usage_output_type
+            and self._n % self._memory_usage_output_n != 0
+        ):
             mem_usage.append(process.memory_info().rss / 1024**2)
-            mem_usage_n.append(n)
+            mem_usage_n.append(self._n)
 
         # Final adaptivity metrics logging at the end of the simulation if not already logged at the end of the last time window
-        if self._is_adaptivity_on and n % self._adaptivity_output_n != 0:
-            self._adaptivity_controller.log_metrics(n)
+        if self._is_adaptivity_on and self._n % self._adaptivity_output_n != 0:
+            self._adaptivity_controller.log_metrics(self._n)
 
         if (
             self._memory_usage_output_type == "all"
@@ -510,18 +528,30 @@ class MicroManagerCoupling(MicroManager):
         if self._interpolate_crashed_sims:
             self._interpolant = Interpolation(self._logger)
 
-        micro_problem = getattr(
-            importlib.import_module(
-                self._config.get_micro_file_name(), "MicroSimulation"
-            ),
-            "MicroSimulation",
-        )
+        micro_problem_cls = None
+        if self._is_model_adaptivity_on:
+            self._model_adaptivity_controller: ModelAdaptivity = ModelAdaptivity(
+                self._config, self._rank, self._log_file
+            )
+            micro_problem_cls = (
+                self._model_adaptivity_controller.get_resolution_sim_class(0)
+            )
+        else:
+            micro_problem_base = getattr(
+                importlib.import_module(
+                    self._config.get_micro_file_name(), "MicroSimulation"
+                ),
+                "MicroSimulation",
+            )
+            micro_problem_cls = create_simulation_class(
+                micro_problem_base, "MicroSimulationDefault"
+            )
 
         # Create micro simulation objects
         self._micro_sims = [0] * self._local_number_of_sims
         if not self._lazy_init:
             for i in range(self._local_number_of_sims):
-                self._micro_sims[i] = create_simulation_class(micro_problem)(
+                self._micro_sims[i] = micro_problem_cls(
                     self._global_ids_of_local_sims[i]
                 )
 
@@ -534,6 +564,7 @@ class MicroManagerCoupling(MicroManager):
                         self._logger,
                         self._rank,
                         self._comm,
+                        micro_problem_cls,
                     )
                 )
             elif self._config.get_adaptivity_type() == "global":
@@ -547,6 +578,7 @@ class MicroManagerCoupling(MicroManager):
                             self._logger,
                             self._rank,
                             self._comm,
+                            micro_problem_cls,
                         )
                     )
                 else:
@@ -559,6 +591,7 @@ class MicroManagerCoupling(MicroManager):
                             self._logger,
                             self._rank,
                             self._comm,
+                            micro_problem_cls,
                         )
                     )
 
@@ -600,7 +633,7 @@ class MicroManagerCoupling(MicroManager):
                     return
 
                 for i in active_sim_lids:
-                    self._micro_sims[i] = create_simulation_class(micro_problem)(
+                    self._micro_sims[i] = micro_problem_cls(
                         self._global_ids_of_local_sims[i]
                     )
 
@@ -613,13 +646,13 @@ class MicroManagerCoupling(MicroManager):
         sim_requires_init_data = False
 
         # Check if provided micro simulation has an initialize() method
-        if hasattr(micro_problem, "initialize") and callable(
-            getattr(micro_problem, "initialize")
+        if hasattr(micro_problem_cls, "initialize") and callable(
+            getattr(micro_problem_cls, "initialize")
         ):
             self._micro_sims_init = True  # Starting value before setting
 
             try:  # Try to get the signature of the initialize() method, if it is written in Python
-                argspec = inspect.getfullargspec(micro_problem.initialize)
+                argspec = inspect.getfullargspec(micro_problem_cls.initialize)
                 if (
                     len(argspec.args) == 1
                 ):  # The first argument in the signature is self
@@ -746,8 +779,8 @@ class MicroManagerCoupling(MicroManager):
                             self._micro_sims[i].initialize()
 
         self._micro_sims_have_output = False
-        if hasattr(micro_problem, "output") and callable(
-            getattr(micro_problem, "output")
+        if hasattr(micro_problem_cls, "output") and callable(
+            getattr(micro_problem_cls, "output")
         ):
             self._micro_sims_have_output = True
 
@@ -1002,6 +1035,38 @@ class MicroManagerCoupling(MicroManager):
 
         return micro_sims_output
 
+    def _solve_micro_simulations_with_model_adaptivity(
+        self, micro_sims_input: list, dt: float, solve_variant: Callable
+    ) -> list:
+        self._model_adaptivity_controller.initialise_solve()
+
+        active_sim_ids = None
+        if self._is_adaptivity_on:
+            active_sim_ids = self._adaptivity_controller.get_active_sim_local_ids()
+        output = None
+
+        while self._model_adaptivity_controller.should_iterate():
+            self._model_adaptivity_controller.switch_models(
+                self._mesh_vertex_coords,
+                self._t,
+                micro_sims_input,
+                output,
+                self._micro_sims,
+                active_sim_ids,
+            )
+            output = solve_variant(micro_sims_input, dt)
+            self._model_adaptivity_controller.check_convergence(
+                self._mesh_vertex_coords,
+                self._t,
+                micro_sims_input,
+                output,
+                self._micro_sims,
+                active_sim_ids,
+            )
+
+        self._model_adaptivity_controller.finalise_solve()
+        return output
+
     def _get_solve_variant(self) -> Callable[[list, float], list]:
         """
         Get the solve variant function based on the adaptivity type.
@@ -1016,7 +1081,13 @@ class MicroManagerCoupling(MicroManager):
         else:
             solve_variant = self._solve_micro_simulations
 
-        return solve_variant
+        if self._is_model_adaptivity_on:
+            return partial(
+                self._solve_micro_simulations_with_model_adaptivity,
+                solve_variant=solve_variant,
+            )
+        else:
+            return solve_variant
 
     def _interpolate_output_for_crashed_sim(
         self,