+++ tests/unit/test_adaptivity_parallel.py
@@ -59,6 +59,7 @@ class TestGlobalAdaptivity(TestCase):
             base_logger=MagicMock(),
             rank=self._rank,
             comm=self._comm,
+            micro_problem_cls=MicroSimulation,
         )
 
         adaptivity_controller._is_sim_active = np.array(
@@ -132,6 +133,7 @@ class TestGlobalAdaptivity(TestCase):
             base_logger=MagicMock(),
             rank=self._rank,
             comm=self._comm,
+            micro_problem_cls=MicroSimulation,
         )
 
         adaptivity_controller._adaptivity_data_names = ["data1", "data2"]
@@ -186,6 +188,7 @@ class TestGlobalAdaptivity(TestCase):
             base_logger=MagicMock(),
             rank=self._rank,
             comm=self._comm,
+            micro_problem_cls=MicroSimulation,
         )
 
         adaptivity_controller._is_sim_active = np.array(
@@ -224,6 +227,7 @@ class TestGlobalAdaptivity(TestCase):
             base_logger=MagicMock(),
             rank=self._rank,
             comm=self._comm,
+            micro_problem_cls=MicroSimulation,
         )
 
         actual_ranks_of_sims = adaptivity_controller._get_ranks_of_sims()